software/OpenTagServer/src/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
redis==6.0.0
requests==2.32.3
pycryptodomex==3.22.0
cryptography==44.0.1
